requirements.txt
jira
openai
python-dotenv
schedule
Flask
Flask-SQLAlchemy
markdown
bleach
